Internal: Refactors interactions handlers [ED-21472]
This improves code organization and maintainability by separating concerns
and promoting code reuse.
