+++ modules/interactions/assets/js/interactions-utils.js
@@ -1,0 +1,49 @@
+
+'use strict';
+
+export const config = window.ElementorInteractionsConfig?.constants || {
+	defaultDuration: 300,
+	defaultDelay: 0,
+	slideDistance: 100,
+	scaleStart: 0.5,
+	easing: 'linear',
+};
+
+export function getKeyframes( effect, type, direction ) {
+	const isIn = 'in' === type;
+	const keyframes = {};
+
+	keyframes.opacity = isIn ? [ 0, 1 ] : [ 1, 0 ];
+
+	if ( 'scale' === effect ) {
+		keyframes.scale = isIn ? [ config.scaleStart, 1 ] : [ 1, config.scaleStart ];
+	}
+
+	if ( direction ) {
+		const distance = config.slideDistance;
+		const movement = {
+			left: { x: isIn ? [ -distance, 0 ] : [ 0, -distance ] },
+			right: { x: isIn ? [ distance, 0 ] : [ 0, distance ] },
+			top: { y: isIn ? [ -distance, 0 ] : [ 0, -distance ] },
+			bottom: { y: isIn ? [ distance, 0 ] : [ 0, distance ] },
+		};
+
+		Object.assign( keyframes, movement[ direction ] );
+	}
+
+	return keyframes;
+}
+
+export function parseAnimationName( name ) {
+	const [ trigger, effect, type, direction, duration, delay ] = name.split( '-' );
+
+	return {
+		trigger,
+		effect,
+		type,
+		direction: direction || null,
+		duration: duration ? parseInt( duration, 10 ) : config.defaultDuration,
+		delay: delay ? parseInt( delay, 10 ) : config.defaultDelay,
+	};
+}
+
